Moved and updated the particle and a transformator function
