+++ propulate/particle.py
